Customer_Management/requirements.txt
Django==4.2 
djangorestframework==3.14.0  
django-cors-headers==3.14.0  
pyjwt==2.9.0